pyproject.toml
[project]
name = "ferdowsi-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4.0",
    "django>=5.1.6",
    "graphene-django>=3.2.2",
    "psycopg[binary]>=3.2.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
]
